+++ server/controllers/authController.js
@@ -1,7 +1,9 @@
 import axios from 'axios';
 import User from '../models/User.js';
 import qs from 'qs';
-import dotenv from 'dotenv';
+import dotenv from 'dotenv'
+import NoobsRecentFriend from "../models/Noobs_Recent_Friend.js";
+
 
 dotenv.config();
 
@@ -47,17 +49,15 @@ const kakaoLogin = async (req, res) => {
       },
     });
 
+    
     // 세션에 사용자 정보 저장
-    req.session.user = {
-      id : user.id,
-      nickname: user.nickname,
-      profileImage: user.profileImage,
-    };
-
-    console.log("세션 저장:", req.session.user);
-
-    // 로그인 후, 세션을 저장하고 리다이렉트
+    req.session.userId = user.id;
+    req.session.nickname = user.nickname;
+    // 세션 저장 후 리다이렉트
     res.redirect('http://localhost:5173/main');
+
+    console.log('카카오로그인', req.session);
+
   } catch (error) {
     console.error(error);
     res.status(500).send("카카오 로그인 실패");
@@ -75,4 +75,55 @@ const logout = (req, res) => {
   });
 };
 
-export { kakaoLogin, logout };
+// 같이 한 사용자 추가 로직
+const userAdd = async (req,res) => {
+ 
+  const { userid, tagLine } = req.query;  
+ 
+
+  console.log('세션 : ' , req.session);
+
+  if (!userid || !tagLine) {
+      return res.status(400).json({ message: '소환사 명을 입력하세요' });
+  }
+
+  // 해당 사용자 db에서 검색
+  try {
+      // DB에서 사용자 검색
+      const userSearchData = await NoobsUserInfo.findOne({
+          where: {
+              gameName: userid,
+              tagLine: tagLine,
+          }
+      });
+
+      if (!userSearchData) {
+          res.status(404).json( { message : '해당 사용자를 찾을 수 없습니다. '});
+      } else {
+
+      // DB 저장: 사용자 정보
+      const user = await NoobsRecentFriend.create({
+          user_id: user_id,  // 세션에서 가져온 user_id 값
+          gameName: userSearchData.gameName,
+          tagLine: userSearchData.tagLine,
+          profileIconId: userSearchData.profileIconId,
+          tier: userSearchData.tier,
+          rank: userSearchData.rank,
+          wins: userSearchData.wins,
+          losses: userSearchData.losses,
+          winRate: userSearchData.winRate,
+      });
+          return res.status(200).json({ userSearchData });
+      }
+
+  } catch (error) {
+      console.error('DB 처리 중 에러 발생:', error);
+      return res.status(500).json({ message: '서버 오류' });
+  }
+  
+};
+
+
+
+
+export { kakaoLogin, logout, userAdd };
