+++ server/controllers/userInfoController.js
@@ -2,7 +2,6 @@ import axios from "axios";
 import dotenv from 'dotenv';
 import NoobsUserInfo from "../models/Noobs_user_info.js";
 import NoobsMasterChamp from "../models/Noobs_master_champ.js";
-import NoobsRecentFriend from "../models/Noobs_Recent_Friend.js";
 
 dotenv.config(); // .env 파일 로드
 
@@ -126,52 +125,5 @@ const userSearch = async (req, res) => {
     }
 };
 
-// 같이 한 사용자 추가 로직
-const userAdd = async (req,res) => {
-    const { userid, tagLine } = req.query;  
-   
 
-    console.log('세션 : ' , req.session);
-
-    if (!userid || !tagLine) {
-        return res.status(400).json({ message: '소환사 명을 입력하세요' });
-    }
-
-    // 해당 사용자 db에서 검색
-    try {
-        // DB에서 사용자 검색
-        const userSearchData = await NoobsUserInfo.findOne({
-            where: {
-                gameName: userid,
-                tagLine: tagLine,
-            }
-        });
-
-        if (!userSearchData) {
-            res.status(404).json( { message : '해당 사용자를 찾을 수 없습니다. '});
-        } else {
-
-        // DB 저장: 사용자 정보
-        const user = await NoobsRecentFriend.create({
-            user_id: user_id,  // 세션에서 가져온 user_id 값
-            gameName: userSearchData.gameName,
-            tagLine: userSearchData.tagLine,
-            profileIconId: userSearchData.profileIconId,
-            tier: userSearchData.tier,
-            rank: userSearchData.rank,
-            wins: userSearchData.wins,
-            losses: userSearchData.losses,
-            winRate: userSearchData.winRate,
-        });
-            return res.status(200).json({ userSearchData });
-        }
-
-    } catch (error) {
-        console.error('DB 처리 중 에러 발생:', error);
-        return res.status(500).json({ message: '서버 오류' });
-    }
-    
-};
-
-
-export { userSearch, userAdd };
+export { userSearch };
