+++ server/routes/authRoutes.js
@@ -1,5 +1,6 @@
 import express from 'express';
-import { kakaoLogin, logout } from '../controllers/authController.js';
+import { kakaoLogin, logout,userAdd } from '../controllers/authController.js';
+
 
 const router = express.Router();
 
@@ -8,5 +9,8 @@ router.get('/user/kakao/login', kakaoLogin);
 
 // 로그아웃
 router.get('/logout', logout);
+
+// 유저추가 [ 최근에 같이한 플레이어 ]
+router.post('/noobs/lolUserAdd', userAdd);
 
 export default router;
