+++ server/routes/route.js
@@ -1,6 +1,6 @@
 import express from 'express';
 import { ChampionUpdate, profileUpdate } from  '../controllers/updateController.js'; // 확장자 .js 추가
-import { userSearch, userAdd } from '../controllers/userInfoController.js'
+import { userSearch } from '../controllers/userInfoController.js'
 
 const router = express.Router();
 
@@ -13,6 +13,6 @@ router.post('/noobs/profileUpadate', profileUpdate);
 router.get('/noobs/lolUser', userSearch);
 
 // 유저추가 [ 최근에 같이한 플레이어 ]
-router.post('/noobs/lolUserAdd', userAdd)
+// router.post('/noobs/lolUserAdd', userAdd);
 
 export default router; // ESM 방식으로 export
