+++ server/server.js
@@ -14,6 +14,16 @@ import cors from "cors"; // cors 설정 추가 했습니다.
 const app = express();
 const PORT = process.env.PORT || 8000;
 
+// CORS 설정
+app.use(
+  cors({
+    origin: "http://localhost:5173/", // 요청을 허용할 클라이언트 도메인
+    methods: ["GET", "POST", "UPDATE", "PATCH"], // 허용할 HTTP 메서드
+    credentials: true, // 쿠키를 허용
+  })
+);
+
+
 // 세션 설정
 app.use(
   session({
@@ -21,19 +31,13 @@ app.use(
     resave: false,
     saveUninitialized: true,
     cookie: {
+      httpOnly: true,
       secure: false, // 개발 중에는 false로 설정 (HTTPS에서만 true)
       maxAge: 10 * 60 * 1000, // 세션 만료 10분
     },
   })
 );
 
-// CORS 설정
-app.use(
-  cors({
-    origin: "http://localhost:5173", // 요청을 허용할 클라이언트 도메인
-    methods: ["GET", "POST", "UPDATE", "PATCH"], // 허용할 HTTP 메서드
-  })
-);
 
 // 현재 파일의 경로를 __dirname처럼 사용하기
 const __filename = fileURLToPath(import.meta.url);
